+++ src/interfaces/ImageInfo.ts
@@ -1,0 +1,5 @@
+import gm from 'gm';
+
+export default interface ImageInfo extends gm.ImageInfo {
+  Scene?: string[]
+}