+++ types/interfaces/ImageInfo.d.ts
@@ -1,0 +1,4 @@
+import gm from 'gm';
+export default interface ImageInfo extends gm.ImageInfo {
+    Scene?: string[];
+}